Revert "Feature/implement simple cov"
