--- .rspec
@@ -1,1 +1,0 @@
---require spec_helper
