--- Gemfile
@@ -1,16 +1,0 @@
-# Gemfile
-#
-
-# Utilise la source globale pour certaines gems
-source 'https://rubygems.org'
-
-# Gems for development and testing
-group :development, :test do
-  gem 'parser'
-  gem 'unparser'
-  gem 'thor'
-  gem 'rspec'
-end
-gem 'simplecov', require: false
-
-
