--- file_with_methods_cp.rb
@@ -1,13 +1,0 @@
-# auinresa pape a totlala
-def trust_no_one
-  puts("Mulder look !")
-end
-
-class Xfiles
-  def trust_no_one
-    puts("Jane look !")
-  end
-end
-
-class Bfiles
-end
