--- spec/fixtures/a_basic_fixture.rb
@@ -1,4 +1,0 @@
-line 1
-line 2
-line 3
-line 4
