--- spec/fixtures/a_basic_fixture_res.rb
@@ -1,3 +1,0 @@
-line 2
-line 3
-line 4
